DAG-438: Legger bundle-logikk i egen hook slik at viewet blir mer oversiktlig
